Update and rename BuildEmail.php to EmailBuilder.php
